修复错误问题

+++ blog-writing/src/main/java/wang.redbean.blog/article/model/entity/ArticleContent.java
@@ -15,7 +15,7 @@ import wang.redbean.blog.base.entity.BaseEntity;
 public class ArticleContent extends BaseEntity {
 
     @TableField("article_id")
-    private Integer articleId;//文章ID
+    private Long articleId;//文章ID
 
     @TableField("content_body")
     private String contentBody;//文章正文
